+++ CloudHarvestApi/api/blueprints/__init__.py
@@ -1,0 +1,2 @@
+from .cache.routes import cache_blueprint
+from .reports.routes import report_blueprint